updated schema to show prices as floats; change js file to show prices with decimal points

+++ schemaBamazon.sql
@@ -7,7 +7,7 @@ CREATE TABLE products (
   item_id INT NOT NULL AUTO_INCREMENT,
   product_name VARCHAR(45) NOT NULL,
   department_name VARCHAR(45) NOT NULL,
-  price INTEGER(10) NULL,
+  price FLOAT(10) NULL,
   stock_quantity  INTEGER(10) NULL,
   PRIMARY KEY (item_id)
 );
